backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
# Force CPU wheels from the PyTorch index
torch==2.1.0+cpu
torchvision==0.16.0+cpu
numpy==1.26.0
scikit-learn==1.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.5.0 
